% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis_pdf_de}
\alias{thesis_pdf_de}
\title{Convert R Markdown to a PDF/\LaTeX or thesis document in German}
\usage{
thesis_pdf_de(
  toc = TRUE,
  toc_depth = 5,
  highlight = "default",
  latex_engine = "pdflatex",
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{toc}{logical; \code{TRUE} to include a table of contents in the output.}

\item{toc_depth}{integer; Depth of headers to include in table of contents. Default set to 5.}

\item{highlight}{character; syntax highlighting style. Supported styles include "default",
"tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock".
Pass \code{NULL} to prevent syntax highlighting.}

\item{latex_engine}{character; \LaTeX engine for producing PDF output. Default is "pdflatex"
as it works well with the German Umlaute.}

\item{pandoc_args}{Additional command line options to pass to pandoc.}

\item{...}{Additional parameters to pass to \code{\link[bookdown]{pdf_book}}.}
}
\value{
A modified \code{\link[rmarkdown]{pdf_document}} based on the UHH LaTeX template.
}
\description{
This function serves as wrapper for \code{\link[bookdown]{pdf_book}}, with different
default values (e.g., \code{highlight = "default"}), a custom Pandoc \LaTeX template
and different knitr default values (e.g., \code{fig.align = "center"}).
It is called from the initial R Markdown template file, which should be named \code{index.Rmd}.
The function is based on the \code{thesis_pdf} function of the
\href{https://github.com/ismayc/thesisdown}{thesisdown} package.
}
\examples{
\dontrun{
 # put in YAML header:
 output: UHHthesis::thesis_pdf_de
}
}
