% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis_word_de}
\alias{thesis_word_de}
\alias{thesis_word_en}
\title{Convert R Markdown to a Word (docx) thesis document in German and English}
\usage{
thesis_word_de(
  toc = TRUE,
  toc_depth = 5,
  number_sections = TRUE,
  highlight = "default",
  reference_docx = "uhh-template.docx",
  pandoc_args = NULL,
  ...
)

thesis_word_en(
  toc = TRUE,
  toc_depth = 5,
  number_sections = TRUE,
  highlight = "default",
  reference_docx = "uhh-template.docx",
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{toc}{logical; \code{TRUE} to include a table of contents in the output.}

\item{toc_depth}{integer; Depth of headers to include in table of contents. Default set to 5.}

\item{number_sections}{logical; whether to number section headers: if TRUE (default), figure/table numbers
will be of the form X.i, where X is the current first-level section number, and i is an
incremental number (the i-th figure/table); if FALSE, figures/tables will be numbered
sequentially in the document from 1, 2, ..., and you cannot cross-reference section
headers in this case.}

\item{highlight}{character; syntax highlighting style. Supported styles include "default",
"tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock".
Pass \code{NULL} to prevent syntax highlighting.}

\item{reference_docx}{character; use the specified file as a style reference in producing a docx file.
The 'uhh-template.docx' template implements most of the standard requirement at the UHH biology
department. If you prefer another template, pass the file name to this argument or simply use
'default' to use your standard Word template.}

\item{pandoc_args}{Additional command line options to pass to pandoc.}

\item{...}{Additional parameters to pass to \code{\link[bookdown]{pdf_book}}.}
}
\value{
A modified \code{\link[rmarkdown]{word_document}} based on the UHH Word template.
}
\description{
These functions serve as wrappers for the bookdown function \code{\link[bookdown]{word_document2}},
with different default values (e.g., \code{number_sections = FALSE}), a custom Pandoc Word template
and different knitr default values (e.g., \code{fig.align = "center"}).
It is called from the initial R Markdown template file, which should be named \code{index.Rmd}.
}
\examples{
\dontrun{
 # put in YAML header:
 output: UHHthesis::thesis_word_de
}
}
