% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHHthesis.R
\docType{package}
\name{UHHthesis}
\alias{UHHthesis}
\alias{UHHthesis-package}
\title{UHHthesis: R Markdown thesis templates designed for the
University of Hamburg (UHH)}
\description{
A package for creating Bachelor and Masters theses
in PDF and MS Word format using bookdown and R Markdown.
}
\section{thesis_pdf_en}{

Creates a PDF/LaTeX thesis document in English
}

\section{thesis_word_en}{

Creates a Microsoft Word (docx) thesis document in English
}

\section{thesis_pdf_de}{

Creates a PDF/LaTeX thesis document inGerman
}

\section{thesis_word_de}{

Creates a Microsoft Word (docx) thesis document in German
}

\author{
\strong{Maintainer}: Saskia Otto \email{saskia.a.otto@gmail.com}

}
